Simulate JSRR R7 according to ISA
Currently, the behavior of `JSRR R7` in LC3Tools does not match Appendix
A of the textbook. It performs the pseudocode:

    R7=PC*
    PC=BaseR

But the LC-3 ISA in the textbook (Appendix A, page 662) actually
specifies the following:

    TEMP=PC*
    PC=BaseR
    R7=TEMP

The behavior is the same _except_ when BaseR=R7. So this commit makes
LC3Tools match the textbook. (Otherwise, the behavior is confusing. I
hit this in the wild and got very confused.)